feat: active-hours window, on-brand range slider, dedicated Pacing section (#185)
- Workweek pacing can now narrow to active hours of the day (same window
  applied to every active day), clipped DST-safe in PacingCalculator.
- HourRangeSlider: on-brand dual-thumb 0-24h control replacing the native
  hour menus, in the TokenEaterSlider visual language.
- Off-time hatch + muted now-marker now honor off-hours too.
- Move pacing logic (Smart Color, sensitivity, workweek) out of Themes
  into a dedicated Pacing settings section; Themes keeps the colour palette.
