+++ Shared/Models/PacingModels.swift
@@ -48,6 +48,12 @@ struct PacingSchedule: Equatable, Sendable {
     var enabled: Bool
     /// Gregorian weekday numbers considered active (1=Sunday ... 7=Saturday).
     var activeDays: Set<Int>
+    /// When true, the active days are further narrowed to `[startHour, endHour)`
+    /// in local time - the off-hours of an active day don't advance the pace
+    /// either. Same hours apply to every active day (no per-day schedules).
+    var hoursEnabled: Bool = false
+    var startHour: Int = 9
+    var endHour: Int = 18
 
     static let allDays: Set<Int> = [1, 2, 3, 4, 5, 6, 7]
     /// Monday through Friday - the default selection when the user enables it.
@@ -63,51 +69,58 @@ struct PacingSchedule: Equatable, Sendable {
         return activeDays
     }
 
-    /// True only when the schedule meaningfully excludes at least one day.
-    /// Drives the workweek badge on pacing surfaces.
-    var isActive: Bool {
-        enabled && !activeDays.isEmpty && activeDays.count < 7
+    /// Active-hours window `(start, end)` in 24h local time, or nil for full
+    /// days. nil whenever the feature is off, hours are off, or the range is
+    /// degenerate (so the calculator never divides by an empty window).
+    var effectiveHours: (start: Int, end: Int)? {
+        guard enabled, hoursEnabled, endHour > startHour else { return nil }
+        return (startHour, endHour)
     }
 
-    /// Whether `date` falls on an excluded (off) day. False unless the schedule
-    /// is active. Drives the "resting" badge variant on off-days.
+    /// True when the schedule meaningfully restricts time - at least one day
+    /// excluded OR the hours narrowed. Drives the workweek badge + off hatch.
+    var isActive: Bool {
+        guard enabled, !activeDays.isEmpty else { return false }
+        return activeDays.count < 7 || effectiveHours != nil
+    }
+
+    /// Whether `date` is outside the active time - an excluded day, or an off
+    /// hour of an active day. False unless active. Drives the "resting" badge
+    /// variant + the muted pace marker.
     func isOffDay(_ date: Date, calendar: Calendar = .current) -> Bool {
         guard isActive else { return false }
-        return !effectiveActiveDays.contains(calendar.component(.weekday, from: date))
+        if !effectiveActiveDays.contains(calendar.component(.weekday, from: date)) { return true }
+        if let h = effectiveHours {
+            let hour = calendar.component(.hour, from: date)
+            if hour < h.start || hour >= h.end { return true }
+        }
+        return false
     }
 
-    /// Off-day spans within the window `[resetDate - period, resetDate]`, as
-    /// x-fractions (0...1) of the calendar window, with contiguous off-days
-    /// merged (a Sat+Sun weekend becomes one band). Empty unless the schedule is
-    /// active. Used to hatch the off zones on the pacing track.
+    /// Off-time spans within the window `[resetDate - period, resetDate]`, as
+    /// x-fractions (0...1), computed as the complement of the active intervals so
+    /// it honors both excluded days and off-hours. Empty unless active.
     func offDayRanges(resetDate: Date, period: TimeInterval = 7 * 24 * 3600, calendar: Calendar = .current) -> [ClosedRange<Double>] {
         guard isActive else { return [] }
         let windowStart = resetDate.addingTimeInterval(-period)
+        let active = PacingCalculator.activeIntervals(
+            from: windowStart, to: resetDate,
+            activeDays: effectiveActiveDays, hours: effectiveHours, calendar: calendar
+        )
+        func frac(_ d: Date) -> Double { min(max(d.timeIntervalSince(windowStart) / period, 0), 1) }
         var ranges: [ClosedRange<Double>] = []
-        var current: (start: Double, end: Double)?
-        var cursor = windowStart
-        var guardCount = 0
-        while cursor < resetDate && guardCount < 400 {
-            guardCount += 1
-            let dayStart = calendar.startOfDay(for: cursor)
-            let nextMidnight = calendar.date(byAdding: .day, value: 1, to: dayStart) ?? resetDate
-            let segEnd = min(nextMidnight, resetDate)
-            let s = cursor.timeIntervalSince(windowStart) / period
-            let e = segEnd.timeIntervalSince(windowStart) / period
-            if !effectiveActiveDays.contains(calendar.component(.weekday, from: cursor)) {
-                if let cur = current, abs(cur.end - s) < 0.0001 {
-                    current = (cur.start, e)
-                } else {
-                    if let cur = current { ranges.append(cur.start...cur.end) }
-                    current = (s, e)
-                }
-            } else if let cur = current {
-                ranges.append(cur.start...cur.end)
-                current = nil
+        var prevEnd = windowStart
+        for iv in active {
+            if iv.start > prevEnd {
+                let lo = frac(prevEnd), hi = frac(iv.start)
+                if hi > lo { ranges.append(lo...hi) }
             }
-            cursor = segEnd
+            if iv.end > prevEnd { prevEnd = iv.end }
         }
-        if let cur = current { ranges.append(cur.start...cur.end) }
+        if prevEnd < resetDate {
+            let lo = frac(prevEnd), hi = frac(resetDate)
+            if hi > lo { ranges.append(lo...hi) }
+        }
         return ranges
     }
 }
